--- packages/engine/src/datastore/arrow/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-mod test;
