+++ packages/engine/src/lib.rs
@@ -42,4 +42,7 @@ pub mod proto;
 pub mod simulation;
 pub mod utils;
 
+#[cfg(test)]
+mod tests;
+
 pub use self::args::{args, Args};
