+++ packages/engine/src/tests/mod.rs
@@ -1,0 +1,6 @@
+pub mod test_utils;
+
+mod agent;
+mod arrow;
+mod datastore;
+mod schema;
